logical_dns: moving to an extension
