+++ CODEOWNERS
@@ -280,6 +280,7 @@ extensions/filters/http/oauth2 @derekargueta @snowp
 /*/extensions/filters/network/common @UNOWNED @UNOWNED
 /*/extensions/clusters/original_dst @UNOWNED @UNOWNED
 /*/extensions/clusters/logical_dns/ @UNOWNED @UNOWNED
+/*/extensions/clusters/common/ @UNOWNED @UNOWNED
 
 # URL Pattern Match and Rewrite Library
 /*/extensions/path/uri_template_lib @alyssawilk @yanjunxiang-google
