+++ source/extensions/clusters/common/logical_host.cc
@@ -1,4 +1,4 @@
-#include "source/extensions/clusters/logical_dns/logical_host.h"
+#include "source/extensions/clusters/common/logical_host.h"
 
 namespace Envoy {
 namespace Upstream {
