+++ source/extensions/clusters/dynamic_forward_proxy/BUILD
@@ -15,7 +15,7 @@ envoy_cc_extension(
     deps = [
         "//source/common/network:transport_socket_options_lib",
         "//source/common/upstream:cluster_factory_lib",
-        "//source/common/upstream:logical_host_lib",
+        "//source/extensions/clusters/common:logical_host_lib",
         "//source/extensions/common/dynamic_forward_proxy:dns_cache_interface",
         "//source/extensions/common/dynamic_forward_proxy:dns_cache_manager_impl",
         "//source/extensions/transport_sockets/tls/cert_validator:cert_validator_lib",
